+++ packages/react-native-sdk/src/components/ActiveCall.tsx
@@ -12,6 +12,7 @@ import { useIncallManager } from '../hooks/useIncallManager';
 import { usePublishMediaStreams } from '../hooks/usePublishMediaStreams';
 import { CallParticipantsBadge } from './CallParticipantsBadge';
 import { PermissionRequests } from './PermissionRequests';
+import { verifyAndroidBluetoothPermissions } from '../utils/verifyAndroidBluetoothPermissions';
 
 /**
  * Props to be passed for the ActiveCall component.
@@ -37,6 +38,8 @@ export const ActiveCall = (props: ActiveCallProps) => {
   activeCallRef.current = activeCall;
 
   useEffect(() => {
+    // when the component mounts, we ask for necessary permissions.
+    verifyAndroidBluetoothPermissions();
     return () => {
       // ensure that if this component is unmounted, the call is left.
       activeCallRef.current?.leave();