+++ packages/react-native-sdk/src/components/index.ts
@@ -9,6 +9,6 @@ export * from './CallParticipantsView';
 export * from './CallControlsView';
 export * from './CallParticipantsInfoView';
 export * from './LobbyView';
-export * from './ParticipantReactions';
 export * from './PermissionsNotification';
+export * from './ParticipantReaction';
 export * from './ReactionsModal';