+++ packages/react-sdk/src/components/CallControls/ToggleAudioButton.tsx
@@ -3,15 +3,15 @@ import { OwnCapability, SfuModels } from '@stream-io/video-client';
 import {
   useCall,
   useHasPermissions,
-  useLocalParticipant,
   useI18n,
+  CallPermissionsWrapper,
+  useLocalParticipant,
 } from '@stream-io/video-react-bindings';
 
-import { useMediaDevices } from '../../core/contexts';
+import { useMediaDevices } from '../../core';
 import { DeviceSelectorAudioInput } from '../DeviceSettings';
 import { CompositeButton, IconButton } from '../Button';
 import { PermissionNotification } from '../Notification';
-import { Restricted } from '../Moderation';
 
 export type ToggleAudioPreviewButtonProps = { caption?: string };
 
@@ -39,12 +39,15 @@ export type ToggleAudioPublishingButtonProps = {
   caption?: string;
 };
 
-export const ToggleAudioPublishingButton = ({
-  caption,
-}: ToggleAudioPublishingButtonProps) => {
+export const ToggleAudioPublishingButton = (
+  props: ToggleAudioPublishingButtonProps,
+) => {
   const { publishAudioStream, stopPublishingAudio } = useMediaDevices();
   const localParticipant = useLocalParticipant();
   const { t } = useI18n();
+
+  const { caption = t('Mic') } = props;
+
   const isAudioMute = !localParticipant?.publishedTracks.includes(
     SfuModels.TrackType.AUDIO,
   );
@@ -88,7 +91,7 @@ export const ToggleAudioPublishingButton = ({
   ]);
 
   return (
-    <Restricted requiredGrants={[OwnCapability.SEND_AUDIO]}>
+    <CallPermissionsWrapper requiredGrants={[OwnCapability.SEND_AUDIO]}>
       <PermissionNotification
         permission={OwnCapability.SEND_AUDIO}
         isAwaitingApproval={isAwaitingApproval}
@@ -99,7 +102,7 @@ export const ToggleAudioPublishingButton = ({
         <CompositeButton
           Menu={DeviceSelectorAudioInput}
           active={isAudioMute}
-          caption={caption || t('Mic')}
+          caption={caption}
         >
           <IconButton
             icon={isAudioMute ? 'mic-off' : 'mic'}
@@ -107,6 +110,6 @@ export const ToggleAudioPublishingButton = ({
           />
         </CompositeButton>
       </PermissionNotification>
-    </Restricted>
+    </CallPermissionsWrapper>
   );
 };