+++ packages/client/CHANGELOG.md
@@ -2,6 +2,50 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.137](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.136...client0.0.1-alpha.137) (2023-05-12)
+
+
+
+## [0.0.1-alpha.136](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.135...client0.0.1-alpha.136) (2023-05-12)
+
+
+### Features
+
+* Livestream sample app ([#489](https://github.com/GetStream/stream-video-js/issues/489)) ([bf6e6e5](https://github.com/GetStream/stream-video-js/commit/bf6e6e54dab884828ca08208f25b1285cf3f1944))
+
+
+
+## [0.0.1-alpha.135](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.134...client0.0.1-alpha.135) (2023-05-12)
+
+
+### Bug Fixes
+
+* disable sdp munging ([#491](https://github.com/GetStream/stream-video-js/issues/491)) ([5bca195](https://github.com/GetStream/stream-video-js/commit/5bca195dbb99efb77c5d7106a251e88d9a045498)), closes [#395](https://github.com/GetStream/stream-video-js/issues/395)
+
+
+### Features
+
+* enable codec preferring through sdp munging ([#395](https://github.com/GetStream/stream-video-js/issues/395)) ([a5db5a8](https://github.com/GetStream/stream-video-js/commit/a5db5a831ccfc833ecfc29e22540be0f0121287d))
+
+
+
+## [0.0.1-alpha.134](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.133...client0.0.1-alpha.134) (2023-05-11)
+
+
+### Features
+
+* server-side participant counting ([#481](https://github.com/GetStream/stream-video-js/issues/481)) ([3410c25](https://github.com/GetStream/stream-video-js/commit/3410c25ec1449f4d39b44080ad64238b38446612))
+
+
+
+## [0.0.1-alpha.133](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.132...client0.0.1-alpha.133) (2023-05-11)
+
+
+
+## [0.0.1-alpha.132](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.131...client0.0.1-alpha.132) (2023-05-11)
+
+
+
 ## [0.0.1-alpha.131](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.130...client0.0.1-alpha.131) (2023-05-10)